% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_and_helper_functions.R
\name{calculate_change}
\alias{calculate_change}
\title{Calculate period-over-period changes for specified columns}
\usage{
calculate_change(
  data,
  analysis_vars,
  weight_var,
  est_type,
  lag_period,
  is_percent,
  calculate_percent = TRUE
)
}
\arguments{
\item{data}{Data frame containing time series data}

\item{analysis_vars}{Character vector of grouping variables}

\item{lag_period}{Number of periods to lag}

\item{calculate_percent}{Whether to calculate percent change}
}
\value{
Data frame with added change columns
}
\description{
Calculate period-over-period changes for specified columns
}
