% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/determine_summaries_to_calculate_on_monthly_file.R
\name{calculate_distribution_summaries}
\alias{calculate_distribution_summaries}
\title{Calculate summaries for a distribution estimate (iterating over levels)}
\usage{
calculate_distribution_summaries(
  month_data,
  estimate,
  current_date,
  lfs_config_list,
  required_months
)
}
\arguments{
\item{month_data}{Data for the month}

\item{estimate}{Estimate configuration}

\item{current_date}{Current date}

\item{lfs_config_list}{Overall configuration}

\item{required_months}{All required months}
}
\value{
Data frame with stacked summaries for all levels
}
\description{
Calculate summaries for a distribution estimate (iterating over levels)
}
\keyword{internal}
