% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_and_helper_functions.R
\name{calculate_marginal_totals}
\alias{calculate_marginal_totals}
\title{Calculate marginal totals}
\usage{
calculate_marginal_totals(data, analysis_vars, est_config = NULL)
}
\arguments{
\item{data}{A data frame containing the data to analyze}

\item{analysis_vars}{A character vector of column names to use for marginal totals}
}
\value{
A data frame with the original data plus additional rows for marginal totals
}
\description{
Calculate marginal totals
}
