% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_and_helper_functions.R
\name{clean_final_results}
\alias{clean_final_results}
\title{Clean final results}
\usage{
clean_final_results(data)
}
\arguments{
\item{data}{Data frame to clean}
}
\value{
Cleaned data frame
}
\description{
Adds some final touches, such as reordering columns to make sure DATE is first, as well as deleting any unrounded raw data
}
