% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_and_helper_functions.R
\name{complete_combinations}
\alias{complete_combinations}
\title{Complete Dataframe Combinations}
\usage{
complete_combinations(data, analysis_vars, weight_var)
}
\description{
Ensures that a dataframe contains rows for all combinations of DATE
and specified analysis variables within the range of the data.
Implicitly missing combinations are made explicit by adding rows
where specified value columns are filled, typically with 0.
This is crucial before performing operations like moving averages,
lag calculations, or marginal totals on grouped time series data.
}
