% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_folder_paths_and_install_dvs.R
\name{create_custom_dvs_file}
\alias{create_custom_dvs_file}
\title{Create custom derived variables file}
\usage{
create_custom_dvs_file()
}
\value{
None
}
\description{
Checks for the existence of the custom derived variables file and creates it
if it doesn't exist. The file includes default transformation logic for commonly
used derived variables.
}
