% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_folder_paths_and_install_dvs.R
\docType{data}
\name{custom_dvs_file}
\alias{custom_dvs_file}
\title{Manage custom derived variables for LFS data}
\format{
An object of class \code{character} of length 1.
}
\usage{
custom_dvs_file
}
\description{
This file provides functionality to manage custom derived variables
for Labour Force Survey data processing.
}
\details{
The functions in this file enable users to:
\itemize{
\item Create a custom_dvs.R file in the user's documents if it doesn't exist
\item Load custom derived variables defined in the custom_dvs.R file
\item Open the custom_dvs.R file for editing
}
}
\keyword{internal}
