% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/determine_summaries_to_calculate_on_monthly_file.R
\name{determine_summaries_to_generate}
\alias{determine_summaries_to_generate}
\title{Determine which summaries to generate}
\usage{
determine_summaries_to_generate(lfs_config_list)
}
\arguments{
\item{lfs_config_list}{Configuration list}
}
\value{
List of summarise expressions
}
\description{
Determine which summaries to generate
}
\keyword{internal}
