% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_folder_paths_and_install_dvs.R
\name{get_config}
\alias{get_config}
\title{Get a configuration value}
\usage{
get_config(name)
}
\arguments{
\item{name}{Name of the configuration value to retrieve}
}
\value{
The configuration value
}
\description{
Get a configuration value
}
