% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_and_helper_functions.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Default value if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Null coalescing operator
}
\keyword{internal}
