% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_folder_paths_and_install_dvs.R
\name{load_custom_dvs}
\alias{load_custom_dvs}
\title{Load custom derived variables}
\usage{
load_custom_dvs()
}
\value{
None
}
\description{
Loads the custom derived variables from the \code{custom_dvs.R} file.
If the file is missing, it will be created automatically.
}
