% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_files_per_month.R
\name{load_month_data}
\alias{load_month_data}
\title{Load data for a single month}
\usage{
load_month_data(current_date, lfs_config_list)
}
\arguments{
\item{current_date}{Date to process}

\item{lfs_config_list}{Configuration list}
}
\value{
Data frame with loaded data
}
\description{
Load data for a single month
}
\keyword{internal}
