% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_to_read_each_raw_file_type.R
\name{load_north_data}
\alias{load_north_data}
\title{Load NORTH data for a specific month}
\usage{
load_north_data(current_date, use_prerelease = FALSE, lfs_config_list)
}
\arguments{
\item{current_date}{Date to process}

\item{use_prerelease}{Whether to use pre-release data}

\item{lfs_config_list}{Configuration list}
}
\value{
Data frame with NORTH data or NULL if not found
}
\description{
Load NORTH data for a specific month
}
\keyword{internal}
