% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_to_read_each_raw_file_type.R
\name{load_plus_north_data}
\alias{load_plus_north_data}
\title{Load TABS Plus North data for a specific month}
\usage{
load_plus_north_data(current_date, use_prerelease = FALSE)
}
\arguments{
\item{current_date}{Date to process}

\item{use_prerelease}{Whether to use pre-release data}
}
\value{
Data frame with TABS Plus data or NULL if not found
}
\description{
Load TABS Plus North data for a specific month
}
\keyword{internal}
