% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_to_read_each_raw_file_type.R
\name{load_quarterly_supplement_data}
\alias{load_quarterly_supplement_data}
\title{Load Quarterly SUPPLEMENT (LMSI) data for a specific month}
\usage{
load_quarterly_supplement_data(current_date, lfs_config_list)
}
\arguments{
\item{current_date}{Date to process}
}
\value{
Data frame with SUPPLEMENT data or NULL if not found
}
\description{
Load Quarterly SUPPLEMENT (LMSI) data for a specific month
}
\keyword{internal}
