% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_to_read_each_raw_file_type.R
\name{load_supplement_bootstraps}
\alias{load_supplement_bootstraps}
\title{Load supplement Bootstrap data for a specific month}
\usage{
load_supplement_bootstraps(
  df_main,
  current_date,
  base_path,
  base_name,
  lfs_config_list
)
}
\arguments{
\item{df_main}{Main supplement data}

\item{current_date}{Date to process}

\item{base_path}{Folder path of supplement data}

\item{base_name}{Filename of main supplement data}
}
\value{
Data frame with SUPPLEMENT data joined with Bootstrap weights or NULL if not found
}
\description{
Load supplement Bootstrap data for a specific month
}
\keyword{internal}
