% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_to_read_each_raw_file_type.R
\name{load_supplement_data}
\alias{load_supplement_data}
\title{Load Monthly SUPPLEMENT (LMI) data for a specific month}
\usage{
load_supplement_data(current_date, use_prerelease = FALSE, lfs_config_list)
}
\arguments{
\item{current_date}{Date to process}

\item{use_prerelease}{Whether to use pre-release data}
}
\value{
Data frame with SUPPLEMENT data or NULL if not found
}
\description{
Load Monthly SUPPLEMENT (LMI) data for a specific month
}
\keyword{internal}
