% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_to_read_each_raw_file_type.R
\name{load_tabs_data}
\alias{load_tabs_data}
\title{Load TABS Data}
\usage{
load_tabs_data(current_date, use_prerelease = FALSE, lfs_config_list)
}
\arguments{
\item{current_date}{Date to process}

\item{use_prerelease}{Whether to use pre-release data}

\item{lfs_config_list}{Configuration list}
}
\value{
Data frame with TABS data
}
\description{
Loads TABS data for a specific month from the appropriate data source.
}
\keyword{Load}
\keyword{TABS}
\keyword{a}
\keyword{data}
\keyword{for}
\keyword{internal}
\keyword{month}
\keyword{specific}
