% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_folder_paths_and_install_dvs.R
\name{open_custom_dvs}
\alias{open_custom_dvs}
\title{Open custom derived variables file}
\usage{
open_custom_dvs()
}
\value{
None
}
\description{
Opens the \code{custom_dvs.R} file in the user's default text editor.
If running in RStudio, it will open in the RStudio source editor.
If the file doesn't exist, it will be created with default content.
}
