% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_microdata.R
\name{process_microdata}
\alias{process_microdata}
\title{Process microdata without estimates}
\usage{
process_microdata(lfs_config_list, required_months)
}
\arguments{
\item{lfs_config_list}{Configuration}
}
\value{
Data frame with microdata
}
\description{
Process microdata without estimates
}
\keyword{internal}
